C51-278: Calculate Total Duration of a Case
To calculate the total duration of a case in an open status, we needed to add
a case status log, to store all the consequent openings and closings a case
can go through in its life-cycle. This log is stored on a new table, which
stores the case's ID, start date, start activity, end date, end activity and
total duration. To calculate total duration, we add the total duration of each
open-close cycle and store this information in a new custom field created for
this purpose.

Conflicts:
	CRM/Civicase/Upgrader.php
	civicase.php